Fix type hinting and mypy

+++ src/geo_calculator/calculations.py
@@ -1,3 +1,3 @@
-def find_average(list: list) -> int:
+def find_average(list: list) -> float:
     """Returns the avreage of a list"""
     return sum(list) / len(list)
